refactor: move capsule form inside dashboard folder, also change the route to create capsules.
